JWT strategy fix

+++ src/books/books.service.ts
@@ -1,11 +1,17 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
+import { Types } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
-import { Book } from './schemas/book.schema';
+import { Book, BookDocument } from './schemas/book.schema';
 import mongoose from 'mongoose';
 import { Query } from 'express-serve-static-core';
 import { CreateBookDto } from './_utils/dto/requests/create-book.dto';
 import { UpdateBookDto } from './_utils/dto/requests/update-book.dto';
 import { User, UserDocument } from 'src/users/schemas/user.schema';
+import { title } from 'process';
 
 @Injectable()
 export class BooksService {
@@ -14,7 +20,7 @@ export class BooksService {
   ) {}
 
   async findAll(query: Query): Promise<Book[]> {
-    const booksPerPage = 2;
+    const booksPerPage = 10;
     const currentPage = Number(query.page) || 1;
     const skip = booksPerPage * (currentPage - 1);
     const title = query.title
@@ -56,9 +62,13 @@ export class BooksService {
     });
   }
 
-  async deleteById(id: string) {
+  async deleteById(id: string, user: UserDocument) {
     const isValid = mongoose.Types.ObjectId.isValid(id);
     if (!isValid) throw new NotFoundException('Book not found');
+    const book = await this.bookModel.findById(id);
+    if (!book) throw new NotFoundException('Book not found');
+    if (book.user != user)
+      throw new UnauthorizedException('This is not your book');
     return await this.bookModel.findByIdAndDelete(id);
   }
 }
